+++ pwndck/pwndckcli.py
@@ -8,79 +8,15 @@
 
 import argparse
 import fileinput
-import hashlib
 import sys
 import textwrap
 import types
-from importlib.metadata import PackageNotFoundError, version
 from typing import Iterable, List
 
-import requests
-
-from pwndck import FlexiHelpFormatter
-
-apiurl = "https://api.pwnedpasswords.com/range/{}"
-
-try:
-    __version__ = version("pwndck")
-except PackageNotFoundError:
-    __version__ = "?"
-
-
-class PwndException(Exception):
-    pass
-
-
-def get_sha(data: str) -> str:
-    hlib = hashlib.sha1()
-    hlib.update(data.encode("utf-8"))
-    hsh = hlib.hexdigest()
-
-    return hsh.upper()
-
-
-def get_hashes(key: str) -> str:
-    """Return hash-adjacent results
-    per https://haveibeenpwned.com/API/v3#PwnedPasswords"""
-
-    url = apiurl.format(key)
-    headers = {
-        "User-Agent": f"PwndCk/{__version__}",
-        "Add-Padding": "true",
-    }
-    r = requests.get(url, headers=headers)
-    response = r.text
-
-    if r.status_code != 200:
-        raise PwndException(r.reason)
-
-    return response
-
-
-def process_pw(pw: str) -> int:
-    """
-    Returns the number of entries for a password in the Have I Been Pwned
-    database.
-
-    Parameters:
-        pw (str): The password to check.
-
-    Returns:
-        int: The number of entries in the database.
-
-    Raises:
-        PwndException: For web query errors.
-    """
-    hsh = get_sha(pw)
-    key = hsh[0:5]
-    body = hsh[5:]
-
-    for line in get_hashes(key).splitlines():
-        if line.startswith(body):
-            (body, count) = line.split(":")
-            return int(count)
-
-    return 0
+from pwndck.db_size import estimate_db, fmt_num
+from pwndck.flexi_formatter import FlexiHelpFormatter
+from pwndck.processpw import PwndException, process_pw
+from pwndck.version import __version__
 
 
 def parse_args():
@@ -139,6 +75,13 @@ def parse_args():
     )
 
     group.add_argument(
+        "-e",
+        "--estimatedb",
+        action="store_true",
+        help="Estimate the current size of the HaveIBeenPwnd password database",
+    )
+
+    group.add_argument(
         "--version", action="version", version=f"%(prog)s {__version__}"
     )
 
@@ -172,6 +115,14 @@ def quiet_print(string, quiet) -> None:
 
 def main() -> None:
     args = parse_args()
+
+    if args.estimatedb:
+        mean, stddev = estimate_db()
+        estimate = fmt_num(mean, 3)
+        print(
+            f"There are currently approximately {estimate} entries in the HaveIBeenPwned password database"
+        )
+        sys.exit(0)
 
     try:
         passwords = get_passwords(args.passwords, args.input)