+++ test/test_pwndckcli.py
@@ -1,0 +1,27 @@
+import pytest
+
+from pwndck.processpw import PwndException
+from pwndck.pwndckcli import get_passwords
+
+
+def test_get_passwords_arg():
+    assert get_passwords(["foo", "bar"], "baz") == ["foo", "bar"]
+
+
+def test_get_passwords_input_file(monkeypatch):
+    monkeypatch.setattr("fileinput.input", lambda files, encoding: ["buzz"])
+    assert [x for x in get_passwords([], "baz")] == ["buzz"]
+
+
+def test_get_passwords_prompt(monkeypatch):
+    monkeypatch.setattr("sys.stdin.isatty", lambda: True)
+    monkeypatch.setattr("builtins.input", lambda prompt: "buzz")
+
+    assert get_passwords([], "") == ["buzz"]
+
+
+def test_get_passwrds_nope(monkeypatch):
+    monkeypatch.setattr("sys.stdin.isatty", lambda: False)
+
+    with pytest.raises(PwndException):
+        get_passwords([], "")